+++ specsim/tests/test_atmosphere.py
@@ -27,8 +27,6 @@ def test_read_only_properties():
     with pytest.raises(AttributeError):
         a.moon = None
     with pytest.raises(AttributeError):
-        a.condition_names = None
-    with pytest.raises(AttributeError):
         a.surface_brightness = None
     with pytest.raises(AttributeError):
         a.extinction = None
@@ -49,8 +47,9 @@ def test_property_updates():
     m = a.moon
 
     assert m._update_required == True
+    sb = 2.0464695e-17
     assert np.allclose(
-        np.mean(a.surface_brightness.value), 1.563611e-17, atol=0.)
+        np.mean(a.surface_brightness.value), sb, atol=0.)
     assert m.visible == True
     assert np.allclose(m.obs_zenith.value, 0.)
     # Evaluating the atmosphere surface_brightness updates the moon.
@@ -63,14 +62,14 @@ def test_property_updates():
     assert m._update_required == True
     assert np.allclose(m.obs_zenith.value, 1.20942920)
     assert np.allclose(
-        np.mean(a.surface_brightness.value), 2.198837e-17, atol=0.)
+        np.mean(a.surface_brightness.value), 6.881726473e-17, atol=0.)
     assert np.allclose(
         np.mean(m.surface_brightness.value), 1.430046e-17, atol=0.)
     assert m._update_required == False
 
     a.airmass = 1.0
     assert np.allclose(
-        np.mean(a.surface_brightness.value), 1.563611e-17, atol=0.)
+        np.mean(a.surface_brightness.value), sb, atol=0.)
     assert np.allclose(
         np.mean(m.surface_brightness.value), 6.370824e-18, atol=0.)
     assert np.allclose(m.obs_zenith.value, 0.)
@@ -103,10 +102,57 @@ def test_seeing_none():
         a.seeing_fwhm_ref = 1.5 * u.arcsec
 
 
+def test_twilight_func():
+    def check(*args):
+        result = twilight_surface_brightness(*args, subtract_dark=None).value
+        assert np.all((result < 21.2) & (result > 18.2))
+    # Check limiting cases.
+    check(0 * u.deg, -12 * u.deg, 0 * u.deg)
+    check(90 * u.deg, -18 * u.deg, 1 * u.deg)
+    # Check broadcasting.
+    check(15 * u.deg, -15 * u.deg, 0 * u.rad)
+    check([15] * u.deg, -15 * u.deg, 0 * u.rad)
+    check([15] * u.deg, [-15] * u.deg, 0 * u.rad)
+    check([15] * u.deg, [-15] * u.deg, [0] * u.rad)
+    check(15 * u.deg, [-15] * u.deg, [0] * u.rad)
+    check(15 * u.deg, [-15] * u.deg, [0] * u.rad)
+    check([15] * u.deg, [-16, -15, -14] * u.deg, 0 * u.rad)
+    check([15, 15] * u.deg, -15 * u.deg, [0, 0] * u.rad)
+    check([15, 15] * u.deg, [-15, -20] * u.deg, [0, 0] * u.rad)
+    check([15, 15] * u.deg, -15 * u.deg, 0 * u.deg)
+    check([[15],[15]] * u.deg, [-16, -15, -14] * u.deg, 0 * u.rad)
+    # Check wraparound in azimuth.
+    check(15 * u.deg, -15 * u.deg, [-400, -90, 0, 90, 180, 270, 400] * u.deg)
+    # Verify range checks.
+    with pytest.raises(ValueError):
+        check(-1 * u.deg, -15 * u.deg, 0 * u.deg)
+    with pytest.raises(ValueError):
+        check([1, 91] * u.deg, -15 * u.deg, 0 * u.deg)
+    with pytest.raises(ValueError):
+        check(15 * u.deg, -10 * u.deg, 0 * u.deg)
+    with pytest.raises(ValueError):
+        check(15 * u.deg, [-10, -15] * u.deg, 0 * u.deg)
+    # Verify broadcasting checks.
+    with pytest.raises(ValueError):
+        check([15, 15] * u.deg, [-15, -15, -15] * u.deg, 0 * u.deg)
+    with pytest.raises(ValueError):
+        check([15, 15] * u.deg, -15 * u.deg, [0, 0, 0] * u.deg)
+    # Verify unit checking.
+    with pytest.raises(ValueError):
+        check(15, -15 * u.deg, 0 * u.deg)
+    with pytest.raises(ValueError):
+        check(15 * u.m, -15 * u.deg, 0 * u.deg)
+
+
 def test_plot():
     c = specsim.config.load_config('test')
     a = initialize(c)
     a.plot()
+
+
+def test_plot_twilight():
+    plot_twilight_brightness(
+        sun_altitude=-15 * u.deg, sun_azimuth=90 * u.deg)
 
 
 def test_plot_moon():